backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.15.0
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
